🎨 improve structure and code quality, ready for PR
