+++ src/SchedulerView/SchedulerViewHeader.js
@@ -2,10 +2,12 @@ import React, { useContext } from 'react'
 import styles from './styles.module.css'
 import { Table } from 'antd'
 import { SchedulerDataContext } from '../index'
+
 const SchedulerViewHeader = ({ days }) => {
   const {
     config: { schedulerViewWidth }
   } = useContext(SchedulerDataContext)
+
   const headerColumns = days.map((date) => ({
     title: date.format('ddd DD/MM'),
     width: schedulerViewWidth / 7
@@ -13,7 +15,7 @@ const SchedulerViewHeader = ({ days }) => {
 
   return (
     <div className={styles.SchedulerViewHeader}>
-      <div style={{ overflow: 'scroll hidden' }}>
+      <div className={styles.SchedulerViewHeaderContainer}>
         <div style={{ width: `${schedulerViewWidth}px` }}>
           <Table
             columns={headerColumns}
