+++ src/SchedulerView/styles.module.css
@@ -1,19 +1,41 @@
+/* Cell */
+.Cell {
+  border-right: 1px solid #f0f0f0;
+}
+
+/* index */
+.ant-table-fixed {
+  table-layout: fixed;
+}
+.SchedulerView {
+  border: 1px solid #f0f0f0;
+  border-right: 0;
+}
+
+/* PopoverContent */
+.PopoverEventTime {
+  text-align: center;
+  font-weight: 500;
+  font-size: 1.5em;
+  margin-top: 0.75em;
+  width: 100%;
+}
+
+/* SchedulerViewBody */
 .ScheduleViewBodyTable table > tbody > tr:hover > td {
   background: #ffffff !important;
 }
 .ScheduleViewBodyTable table > thead > tr > th {
   display: table-cell;
 }
+.SchedulerViewContent {
+  overflow: auto;
+}
 
+/* SchedulerViewHeader */
 .SchedulerViewHeader {
   overflow: hidden;
   height: 46px;
-}
-.PopoverEventTime {
-  text-align: center;
-  font-weight: 500;
-  font-size: 1.5em;
-  margin-top: 0.75em;
 }
 
 .ScheduleViewHeaderTable th {
@@ -22,8 +44,6 @@
   border-bottom: 0 !important;
   border-right: 1px solid #f0f0f0;
 }
-
-.ant-table-fixed {
-  table-layout: fixed;
+.SchedulerViewHeaderContainer {
+  overflow: scroll hidden;
 }
-
